VPN-4501: Update location details dialog (#4062)
