+++ nym-vpn-app/src/constants.ts
@@ -54,5 +54,5 @@ export const NetworkExplorerNodeUrl = 'https://nym.com/explorer/nym-node';
 export const countriesWithRegions = ['US', 'CA', 'AU', 'MX', 'BR', 'IN', 'CN'];
 export const ResidentialIpServersUrl =
   'https://support.nym.com/hc/en-us/articles/35279486714641-Why-can-t-I-access-streaming-services-while-using-NymVPN';
-export const LocationAccuracyLink =
-  'https://support.nym.com/hc/en-us/articles/26448676449297-How-is-server-location-determined-by-NymVPN';
+export const QuicSupportArticleUrl =
+  'https://support.nym.com/hc/en-us/articles/39648047741457-QUIC-transport-mode';
