+++ nym-vpn-app/src/screens/node/location-details-dialog/LocationDetailsDialog.tsx
@@ -1,0 +1,56 @@
+import { DialogTitle } from '@headlessui/react';
+import { useTranslation } from 'react-i18next';
+import { Button, Dialog, MsIcon } from '../../../ui';
+import { capFirst } from '../../../util';
+import { NodeHop } from '../../../types';
+import { Details } from './Details';
+
+export type Props = {
+  isOpen: boolean;
+  onClose: () => void;
+  node: NodeHop;
+};
+
+function LocationDetailsDialog({ isOpen, onClose, node }: Props) {
+  const { t } = useTranslation('nodeLocation');
+  const title =
+    node === 'entry'
+      ? t('location-details.entry-title')
+      : t('location-details.exit-title');
+
+  return (
+    <Dialog
+      open={isOpen}
+      onClose={onClose}
+      className="flex flex-col items-center gap-6"
+      data-testid="location-details-dialog"
+    >
+      <div className="flex flex-col items-center gap-4">
+        <MsIcon
+          icon="info"
+          className="text-3xl text-baltic-sea dark:text-white"
+          data-testid="location-details-info-icon"
+        />
+        <DialogTitle
+          as="h3"
+          className="text-xl text-baltic-sea dark:text-white text-center"
+          data-testid="location-details-title"
+        >
+          {title}
+        </DialogTitle>
+      </div>
+      <Details node={node} />
+      <Button
+        onClick={onClose}
+        className="mt-2"
+        data-testid="location-details-close-button"
+      >
+        <span className="text-lg text-black dark:text-baltic-sea">
+          {capFirst(t('ok', { ns: 'glossary' }))}
+        </span>
+      </Button>
+    </Dialog>
+  );
+}
+
+export default LocationDetailsDialog;
