+++ nym-vpn-app/src/screens/node/location-details-dialog/index.ts
@@ -1,0 +1,1 @@
+export { default as LocationDetailsDialog } from './LocationDetailsDialog';
